ingestion-edge/requirements.txt
blinker==1.4
dockerflow==2018.4.0
Flask==1.0.2
google-cloud-pubsub==0.38.0
gunicorn==19.9.0
persist-queue==0.4.0
pytest-mypy==0.3.2
pytest==3.8.0
pytest-cov==2.6.0
pytest-docstyle==1.4.0
pytest-flake8==1.0.2
python-dateutil==2.7.3
